chore: restructure code - dashboard views - categories
